% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plot Population of a UN Country.}
\usage{
CountryPopulation(Name)
}
\arguments{
\item{Name}{A character string of the country name}
}
\value{
A ggplot object
}
\description{
This function plots population vs year (1950–2020) for a given country.
}
