% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Data (1950–2020)}
\format{
A data frame with rows representing countries and columns:
\describe{
  \item{Country Name}{Name of the country}
  \item{1950–2020}{Population estimates for each year}
}
}
\usage{
WorldPopulation
}
\description{
Population estimates for UN countries from 1950 to 2020.
}
\keyword{datasets}
